demoagent/gameframeworkwithtooldecorator/pyproject.toml
[project]
name = "gameframework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flake8>=7.3.0",
    "litellm>=1.76.0",
    "load-dotenv>=0.1.0",
    "pydantic>=2.11.7",
]
